f-analytics - reply on Store state now ctor is preparing the Store state
