+++ packages/services/f-analytics/src/store/analytics.module.js
@@ -3,12 +3,44 @@ import {
     UPDATE_EVENTS,
     CLEAR_EVENTS
 } from './mutation-types';
-import defaultState from './default-state';
 
 export default {
     namespaced: true,
 
-    state: () => defaultState,
+    state: () => ({
+        platformData: {
+            environment: 'localhost',
+            name: undefined,
+            appType: undefined,
+            applicationId: undefined,
+            userAgent: undefined,
+            branding: undefined,
+            country: undefined,
+            language: undefined,
+            jeUserPercentage: undefined,
+            currency: undefined,
+            version: undefined,
+            instancePosition: undefined,
+            isPilot: undefined
+        },
+        userData: {
+            'a-UserId': undefined,
+            authType: undefined,
+            email: undefined,
+            globalUserId: undefined,
+            signinType: undefined,
+            signupDate: undefined
+        },
+        pageData: {
+            name: undefined,
+            httpStatusCode: 200,
+            conversationId: undefined,
+            requestId: undefined,
+            orientation: undefined,
+            display: undefined
+        },
+        events: []
+    }),
 
     actions: {
         updatePlatformData: ({ commit }, platformData) => {
