+++ packages/services/f-analytics/src/tests/helpers/setup.js
@@ -14,6 +14,41 @@ const newEvent = {
     }
 };
 
+const defaultState = {
+    platformData: {
+        environment: 'localhost',
+        name: undefined,
+        appType: undefined,
+        applicationId: undefined,
+        userAgent: undefined,
+        branding: undefined,
+        country: undefined,
+        language: undefined,
+        jeUserPercentage: undefined,
+        currency: undefined,
+        version: undefined,
+        instancePosition: undefined,
+        isPilot: undefined
+    },
+    userData: {
+        'a-UserId': undefined,
+        authType: undefined,
+        email: undefined,
+        globalUserId: undefined,
+        signinType: undefined,
+        signupDate: undefined
+    },
+    pageData: {
+        name: undefined,
+        httpStatusCode: 200,
+        conversationId: undefined,
+        requestId: undefined,
+        orientation: undefined,
+        display: undefined
+    },
+    events: []
+};
+
 const modifiedState = {
     platformData: {
         environment: 'test-environment',
@@ -98,6 +133,7 @@ const createStore = ({
 };
 
 export {
+    defaultState,
     defaultActions,
     modifiedState,
     newEvent,
